Float

+++ src/aac_test.zig
@@ -23,7 +23,7 @@ test "AAC info" {
     const info = try aac.vtable.info(&br);
     try testing.expectEqual(@as(u32, 44100), info.sample_rate);
     try testing.expectEqual(@as(u8, 2), info.channels);
-    try testing.expectEqual(SampleType.i16, info.sample_type);
+    try testing.expectEqual(SampleType.f32, info.sample_type);
 }
 
 test "AAC decode" {
@@ -32,7 +32,7 @@ test "AAC decode" {
 
     try testing.expectEqual(@as(u32, 44100), audio.params.sample_rate);
     try testing.expectEqual(@as(u8, 2), audio.params.channels);
-    try testing.expectEqual(SampleType.i16, audio.params.sample_type);
+    try testing.expectEqual(SampleType.f32, audio.params.sample_type);
     try testing.expect(audio.data.len > 0);
 }
 
@@ -42,7 +42,7 @@ test "AAC streaming API" {
 
     try testing.expectEqual(@as(u32, 44100), decoder.info.sample_rate);
     try testing.expectEqual(@as(u8, 2), decoder.info.channels);
-    try testing.expectEqual(SampleType.i16, decoder.info.sample_type);
+    try testing.expectEqual(SampleType.f32, decoder.info.sample_type);
 
     var adapter = api.DecoderReader.init(decoder);
     const reader = adapter.reader();
